Moved code to run csv parse/process into data-controller with flag to remove it when FlockWatch-Scraping is live.
