+++ src/controllers/data-controller.ts
@@ -4,6 +4,12 @@ import { USSummaryService } from "../services/model-services/us-summary-service"
 import { logger } from "../utils/winston-logger";
 import { Request, Response } from "express";
 
+//TODO: Remove when FlockWatch-Scraping has been built
+import path from "path";
+import { ReadCSV } from "../utils/csv-parser/read-csv";
+import { CSVParser } from "../utils/csv-parser/csv-parser";
+import { FlockCasesByStateTransformer } from "../utils/csv-parser/transformers/flock-cases-by-state-transformer";
+
 class DataController {
     private flockCasesByStateService: FlockCasesByStateService;
     private lastReportDateService: LastReportDateService;
@@ -48,6 +54,42 @@ class DataController {
                 error: "Failed to fetch last report date!",
             });
         }
+    }
+
+    // TODO: This will be moved to FlockWatch-Scraping which is in charge of the webscrapers and is an entirely separate system, it's here for testing routes at the moment
+    public async updateAllFlockCases(req: Request, res: Response) {
+        try {
+            logger.http(`Received Request at CSV Processing Route ${req.url}`);
+            const csvFilePath = path.resolve(
+                __dirname,
+                "../../data/Map Comparisons.csv"
+            );
+            logger.silly(`Parsing CSV File at ${csvFilePath}`);
+            const csvData = ReadCSV.readCSVFile(csvFilePath, "utf-16le")
+            const customHeaders = [
+                "State Abbreviation",
+                "State Name",
+                "Backyard Flocks",
+                "Birds Affected",
+                "Color",
+                "Commercial Flocks",
+                "Last Reported Detection Text",
+                "Total Flocks",
+                "State Boundary",
+                "State Label",
+                "Latitude (generated)",
+                "Longitude (generated)"
+            ];
+            const parsedData = CSVParser.parseCSV(csvData, "\t", 2, customHeaders)
+            const dataFiltered = parsedData.filter((row: { [x: string]: string }) => row["State Name"]?.trim() && row["Birds Affected"] !== "0");
+            const transformedData = FlockCasesByStateTransformer.transformData(dataFiltered);
+            res.json(transformedData);
+        } catch (error) {
+            logger.error(`Error processing CSV Data: ${error}`)
+            res.status(500).json({
+                error: "Failed to process CSV Data!",
+            });
+        }   
     }
 }
 export { DataController };
