+++ src/services/model-services/flock-cases-by-state-service.ts
@@ -1,4 +1,5 @@
 import { FlockCasesByStateModel } from "../../models/flock-cases-by-state-model";
+
 class FlockCasesByStateService {
     public async getAllFlockCases() {
         return FlockCasesByStateModel.getModel.find({}).select("-_id -__v");
